Remove redundant part of message, and redirect to GET after saving
This change updates the "enter your key" message to not include the
name of the tab that users are already on when they see this message.

It also adds a redirect after the POST request that saves the key
and secret, so that form submission isn't inadvertently repeated.
This also ensures that if the key and secret are still not complete,
the warning message will still be shown.

Lastly, a couple of Flickr URLs in the help text are updated to HTTPS.