Improve import-all functionality
This switches the import-all function to work with any number of
images and without timing out by paging through a user's photos
and importing each photo. The API endpoints are renamed (as a
new one is introduced).

Tag importing is fixed to use the 'raw' value rather than the
spaceless normalized one.

Checksums of files are also checked before import, along with
filenames.

Logging is added throughout the import process.

A button is added to stop the import before it's finished.

Refs: #12